+++ src/SmallCalc/Eval.hs
@@ -6,17 +6,17 @@ where
 import Data.Fixed
 import SmallCalc.AST
 
-binaryOp :: BinaryOp -> (Float -> Float -> Float)
+binaryOp :: BinaryOp -> (Double -> Double -> Double)
 binaryOp Add = (+)
 binaryOp Sub = (-)
 binaryOp Mul = (*)
 binaryOp Div = (/)
 binaryOp Mod = mod'
 
-unaryOp :: UnaryOp -> (Float -> Float)
+unaryOp :: UnaryOp -> (Double -> Double)
 unaryOp Negate = (* (-1.0))
 
-eval :: Node -> Float
+eval :: Node -> Double
 eval (Value (Constant c)) = c
 eval (UnaryOp op v)       = unaryOp op (eval v)
 eval (BinaryOp op lv rv)  = binaryOp op (eval lv) (eval rv)
